Move coredata test fixtures into service testutil
New test fixture interface:\n\nBefore:\n\n    coretesting.NewStubServices(t)\n    coretesting.AttachStubExternalCredentials(services)\n\nAfter:\n\n    testutil.NewStubServices(t)\n    testutil.AttachStubExternalCredentials(services)\n\nExample:\n\n    services := testutil.NewStubServices(t)\n\nThis keeps core/testing limited to pure core provider and storage stubs while the coredata-backed Services fixture lives with service-level test utilities.
