+++ gestaltd/core/testing/stubs.go
@@ -3,32 +3,11 @@ package coretesting
 import (
 	"context"
 	"sync"
-	"testing"
 	"time"
 
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
-	"github.com/valon-technologies/gestalt/server/internal/coredata"
 )
-
-func NewStubServices(t *testing.T) *coredata.Services {
-	t.Helper()
-	svc, err := coredata.New(&StubIndexedDB{})
-	if err != nil {
-		t.Fatalf("NewStubServices: %v", err)
-	}
-	AttachStubExternalCredentials(svc)
-	return svc
-}
-
-func AttachStubExternalCredentials(svc *coredata.Services) *StubExternalCredentialProvider {
-	if svc == nil {
-		return nil
-	}
-	provider := NewStubExternalCredentialProvider()
-	svc.ExternalCredentials = provider
-	return provider
-}
 
 type StubExternalCredentialProvider struct {
 	mu           sync.Mutex
